Keep imported canvas colors in ARGB space
Constraint: Compose Color(ULong) expects packed internal color data, not saved #AARRGGBB values.
Rejected: Catching draw-time exceptions | That would hide invalid color conversion instead of preserving saved color semantics.
Confidence: high
Scope-risk: narrow
Directive: Treat persisted canvas border colors as plain ARGB longs before converting to Compose colors.
Tested: ./gradlew :desktop-app:test --tests dev.agentperf.desktop.CanvasBorderColorsTest; ./gradlew clean test; ./gradlew :shared-kernel:android-agent-view:assembleDebug :samples:android-view-app:assembleDebug :desktop-app:createDistributable
Not-tested: Manual import through the desktop UI.

+++ desktop-viewer/desktop-app/src/main/kotlin/dev/agentperf/desktop/CanvasBorderColors.kt
@@ -6,7 +6,7 @@ import java.util.prefs.Preferences
 
 @JvmInline
 internal value class CanvasArgb(val value: Long) {
-    fun toComposeColor(): Color = Color(value.toULong())
+    fun toComposeColor(): Color = Color(value)
     fun toHex(): String = String.format(Locale.US, "#%08X", value)
 
     companion object {
